fix: Resolve compilation errors in payment history implementation
- Changed CompressedPaymentRecord.status from u8 to u32 (Soroban compatible)
- Fixed import statement formatting (added missing comma)
- Added missing type imports in storage.rs
- Made get_payments_map public for query functions
- Added Display implementations for new error variants
- Moved trait method implementations from trait to impl block

+++ Contracts/payment-processing-contract/src/types.rs
@@ -366,7 +366,7 @@ pub struct CompressedPaymentRecord {
     pub token: Address,
     pub amount: i128,
     pub paid_at: u64,
-    pub status: u8, // 0: paid, 1: partially refunded, 2: fully refunded
+    pub status: u32, // 0: paid, 1: partially refunded, 2: fully refunded
 }
 
 // Event topics
